Update to use full paths to the HSS payload generator
Updated the calls to the payload generator so they use the absolute path to the tool so it can be called from any directory.

Updated the function call to generate a payload (on Linux) to pass the payload source directory to the make_hss_payload function and also pass the absolute path to the payload generator.