Update Linux payload configuration to remove relative path
The HSS payload now being generated from the directory containing the bin file and the HSS payload generator being called using an absolute path.
This means that the payload config file no longer needs a relative path as it can assume the payload will be in the same directory as the payload generator is being called from.