Added native platform warning for Linux
A warning has been added for Linux users to display the softconsole headless help if the script is being run on a native platform.

The native platform check function is now called from the check tool status linux function.
If a native platform is found a warning is printed along with the softconsole headless help and the user is given the option to continue.