Added path length check for Windows
A path length check was added for Windows as it was observed that the script would fail on systems with a long path to the script and long path support disabled.
The check prints out a warning and the current path length so users can determine if they want to continue or how much to reduce thir path length by.